requirements.txt
playwright
python-dotenv
pytest
pytest-playwright
pytest-html
google-generativeai
Pillow
allure-pytest